project FoorballAssistant update after work hardcoded strings issue update

+++ practice/stage_six/FootballAssistant/app/src/main/java/ru/vpcb/footballassistant/utils/Config.java
@@ -487,7 +487,7 @@ public class Config {
 
         if (spanHeight < MIN_HEIGHT) spanHeight = MIN_HEIGHT;
 
-        int minWidth = (int) (MIN_WIDTH);  // horizontal
+        int minWidth = MIN_WIDTH;  // horizontal
         if (spanWidth < minWidth) spanWidth = minWidth;
 
 // vertical
